requirements.txt
slack_sdk==3.40.1
anthropic==0.83.0
beautifulsoup4==4.14.3
requests==2.32.5
python-dotenv==1.2.1
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pg8000==1.31.2
redis==5.0.8
apscheduler==3.10.4
feedparser==6.0.11
jinja2==3.1.5
